🔨Refactor: #17 프론트엔드 홈템플릿 코드 리펙토링
- 기존 HomeTemplate.jsx에 존재했던 리스트출력 코드 부분을 workTimeList.jsx
파일로 옮김
- 사용자의 이름 밑에 총 근무시간 표시하는 기능 추가
